[TweetPublisher] decouple UnpublishedTweetsProvider
